int32

+++ convex_hull.py
@@ -5,7 +5,7 @@ def convex_hull(points1, points2):
     hull1 = []
     hull2 = []
 
-    hullIndex = cv2.convexHull(np.array(points2), returnPoints=False)
+    hullIndex = cv2.convexHull(np.array(points2).astype(np.int32), returnPoints=False)
     # pdb.set_trace()
 
     for i in xrange(0, len(hullIndex)):
@@ -28,7 +28,7 @@ def convex_hull_internal_points(points1, points2, face_landmarks_dict_1,face_lan
 
     # pdb.set_trace()
     print (len(points2))
-    hullIndex = cv2.convexHull(np.array(points2), returnPoints=False)
+    hullIndex = cv2.convexHull(np.array(points2).astype(np.int32), returnPoints=False)
 
     for i in xrange(0, len(hullIndex)):
         hull1.append(points1[int(hullIndex[i])])
